+++ packages/lib/service/attribute/server/utils.ts
@@ -1,4 +1,4 @@
-import { AttributeToUserRepository } from "../../../server/repository/attributeToUser";
+import { AttributeToUserRepository } from "../../../server/repository/attributeToUserRepository";
 
 export const getWhereClauseForAttributeOptionsManagedByCalcom = () => {
   // Neither created nor updated by DSync
