+++ packages/trpc/server/routers/viewer/attributes/list.handler.ts
@@ -1,4 +1,4 @@
-import { AttributeRepository } from "@calcom/lib/server/repository/attribute";
+import { AttributeRepository } from "@calcom/lib/server/repository/attributeRepository";
 
 import { TRPCError } from "@trpc/server";
